backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
pillow==10.1.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
qrcode==7.4.2
requests==2.31.0
django-extensions==3.2.3
whitenoise==6.6.0 